+++ condition_key_equal.go
@@ -1,6 +1,7 @@
 package dasel
 
 import (
+	"fmt"
 	"reflect"
 )
 
@@ -8,6 +9,16 @@ import (
 type KeyEqualCondition struct {
 	// Value is the value we are looking for.
 	Value string
+	// Not is true if this is a not equal check.
+	Not bool
+}
+
+func (c KeyEqualCondition) check(a interface{}, b interface{}) (bool, error) {
+	var res = fmt.Sprint(a) == b
+	if c.Not {
+		res = !res
+	}
+	return res, nil
 }
 
 // Check checks to see if other contains the required key value pair.
@@ -18,5 +29,5 @@ func (c KeyEqualCondition) Check(other reflect.Value) (bool, error) {
 
 	value := unwrapValue(other)
 
-	return c.Value == value.String(), nil
+	return c.check(c.Value, value.String())
 }