+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing
+### Added
+
+- Support for `!=` comparison operator in dynamic and search selectors.
 
 ## [v1.16.1] - 2021-08-02